feat: u8 ply instead of u16

+++ engine/src/search/transposition.rs
@@ -5,7 +5,7 @@ use super::{encoded_move::EncodedMove, eval_data::EvalNumber};
 #[derive(Clone, Copy)]
 pub struct NodeValue {
     pub zobrist_key: Zobrist,
-    pub ply_remaining: u16,
+    pub ply_remaining: u8,
     pub node_type: NodeType,
     pub value: EvalNumber,
     pub transposition_move: EncodedMove,
